% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{diPAP}
\alias{diPAP}
\title{Di-substituted polyfluorinated phosphate esters analysed in all samples}
\format{
A character vector.
}
\description{
Character vector containing the internal identifiers of
di-substituted polyfluorinated phosphate esters (diPAPs) analysed in
all samples.
}
\seealso{
\code{\link{PFAS}}, \code{\link{contam_info}}
}
\keyword{datasets}
