% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{FOSAAs}
\alias{FOSAAs}
\title{Perfluorooctanesulfonamidoacetic acids analysed in all samples}
\format{
A character vector.
}
\description{
Character vector containing the internal identifiers of
perfluorooctanesulfonamidoacetic acids (FOSAAs) analysed in all samples.
}
\seealso{
\code{\link{PFAS}}, \code{\link{contam_info}}
}
\keyword{datasets}
