% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{FOSAs}
\alias{FOSAs}
\title{Perfluorooctanesulfonamides analysed in all samples}
\format{
A character vector.
}
\description{
Character vector containing the internal identifiers of
perfluorooctanesulfonamides (FOSAs) analysed in all samples.
}
\seealso{
\code{\link{FOSAs_ALL}}, \code{\link{PFAS}},
\code{\link{contam_info}}
}
\keyword{datasets}
