% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{FOSAs_ALL}
\alias{FOSAs_ALL}
\title{All perfluorooctanesulfonamides analysed in the CHOPIN project}
\format{
A character vector.
}
\description{
Character vector containing the internal identifiers of all
perfluorooctanesulfonamides (FOSAs) analysed across sediment and biota
samples in the CHOPIN project.
}
\details{
The \code{_ALL} suffix indicates that the vector contains the complete set
analysed across both matrices.
}
\seealso{
\code{\link{FOSAs}},
\code{\link{contam_info}}
}
\keyword{datasets}
