% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{FTSAs}
\alias{FTSAs}
\title{Fluorotelomer sulfonates analysed in biota}
\format{
A character vector.
}
\description{
Character vector containing the internal identifiers of
fluorotelomer sulfonates (FTSAs) analysed in biota samples.
}
\seealso{
\code{\link{PFAS}}, \code{\link{contam_info}}
}
\keyword{datasets}
