% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalised_conc.R
\name{normalised_conc}
\alias{normalised_conc}
\title{Complete data with sum by family-normalised concentrations}
\usage{
normalised_conc(family, contams, data, unit)
}
\arguments{
\item{family}{character family of contaminants}

\item{contams}{R object vector of compounds names}

\item{data}{tibble dataset of concentrations}

\item{unit}{character unit of concentrations}
}
\value{
tibble completed contamination data
}
\description{
Complete data with sum by family-normalised concentrations
}
