% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{other_PFAS}
\alias{other_PFAS}
\title{Other PFAS compounds analysed in all samples}
\format{
A character vector.
}
\description{
Character vector containing the internal identifiers of PFAS compounds
classified outside the main PFAS subfamilies and analysed in all samples.
}
\seealso{
\code{\link{PFAS}}, \code{\link{contam_info}}
}
\keyword{datasets}
