% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{PALETTE_PFAS_benthos}
\alias{PALETTE_PFAS_benthos}
\title{Color palette for PFAS in benthic macro-invertebrates}
\format{
A character vector.

A character vector.
}
\usage{
PALETTE_PFAS_benthos

PALETTE_PFAS_benthos
}
\description{
Character vector containing the colors used to represent
the different PFAS compounds in figures related to benthos.

Character vector containing the colors used to represent
the different PFAS compounds in figures related to benthos.
}
\examples{
PALETTE_PFAS_benthos
PALETTE_PFAS_benthos
}
\keyword{datasets}
