% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{PALETTE_PFAS_sed}
\alias{PALETTE_PFAS_sed}
\title{Color palette for PFAS in sediments}
\format{
A character vector.
}
\usage{
PALETTE_PFAS_sed
}
\description{
Character vector containing the colors used to represent
the different PFAS compounds in figures related to sediments.
}
\examples{
PALETTE_PFAS_sed
}
\keyword{datasets}
