% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{PALETTE_PFAS_soles}
\alias{PALETTE_PFAS_soles}
\title{Color palette for PFAS in juvenile soles}
\format{
A character vector.
}
\usage{
PALETTE_PFAS_soles
}
\description{
Character vector containing the colors used to represent
the different PFAS compounds in figures related to soles.
}
\examples{
PALETTE_PFAS_soles
}
\keyword{datasets}
