% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{PCB}
\alias{PCB}
\title{PCB compounds analysed in all samples}
\format{
A character vector.
}
\description{
Character vector containing the internal identifiers of the
polychlorinated biphenyl (PCB) congeners analysed in all samples.
}
\details{
The values are extracted from \code{contam_info$chemical}.
}
\seealso{
\code{\link{contam_info}}
}
\keyword{datasets}
