% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{PFAS}
\alias{PFAS}
\title{PFAS compounds analysed in all samples}
\format{
A character vector.
}
\description{
Character vector containing the internal identifiers of the
per- and polyfluoroalkyl substances (PFAS) analysed in all samples.
}
\details{
This vector is a matrix-specific subset of the compounds represented in
\code{PFAS_ALL}.
}
\seealso{
\code{\link{PFAS_ALL}},
\code{\link{contam_info}}
}
\keyword{datasets}
