% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{PFSAs}
\alias{PFSAs}
\title{Perfluoroalkyl sulfonates analysed in all samples}
\format{
A character vector.
}
\description{
Character vector containing the internal identifiers of
perfluoroalkyl sulfonates (PFSAs) analysed in all samples.
}
\seealso{
\code{\link{PFAS}}, \code{\link{contam_info}}
}
\keyword{datasets}
