% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{sub_family}
\alias{sub_family}
\title{PFAS subfamilies represented in biota samples}
\format{
A character vector.
}
\description{
Character vector containing the PFAS subfamily assignments for compounds
analysed in biota samples.
}
\seealso{
\code{\link{sub_family_ALL}}, \code{\link{contam_info}}
}
\keyword{datasets}
