% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_by_family.R
\name{sum_by_family}
\alias{sum_by_family}
\title{Complete data with sum of concentrations by family}
\usage{
sum_by_family(family, contams, data, unit)
}
\arguments{
\item{family}{character family of contaminants}

\item{contams}{R object vector of compounds names}

\item{data}{tibble dataset of concentrations}

\item{unit}{character unit of concentrations}
}
\value{
tibble completed contamination data with sum of concentrations by family
}
\description{
Complete data with sum of concentrations by family
}
