% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_conc.R
\name{convert_conc}
\alias{convert_conc}
\title{Compute concentrations in another unit}
\usage{
convert_conc(
  data,
  contams,
  source_unit = NULL,
  target_unit,
  norm_var,
  operation = c("divide", "multiply")
)
}
\arguments{
\item{data}{tibble dataset of source concentrations}

\item{contams}{character vector of contaminant names}

\item{source_unit}{character unit of source concentrations}

\item{target_unit}{character unit of target concentrations}

\item{norm_var}{character suffix of target concentrations}

\item{operation}{character type of operation ("divide", "multiply")}
}
\value{
tibble completed contamination data with target concentrations
}
\description{
Compute concentrations in another unit
}
\examples{
# Add lipid-normalized concentrations for PCBs
benthos_contam <- benthos_contam |>
  convert_conc(
    contams = PCB,
    source_unit = "_ng_g-1ps",
    target_unit = "_ng_g.1pl",
    norm_var = "lip_PS_percent",
    operation = "divide"
  )
}
